Move `_POPTORCH_AVAILABLE` and `_IPU_AVAILABLE` (#16509)

(cherry picked from commit 7bbbe22636853e1ee6fa57b557b3408d45233589)
